Fix the handling of "from:" in ImageStreams and BuildConfigs
In ImageStreams, we only want a `from:` clause (or for that matter, a
`spec:`) in the case the ImageStream is being pulled from an external
Docker registry. This was previously the intent of the code, but it
didn't handle the case where a `from` argument was passed in, but
didn't designate an external registry.

* Refactor to extract _get_from_struct() method

* Call it to interpret the `from` argument in both the BuildConfig and
  ImageStream code paths (in the latter, replacing the ad-hoc and
  buggy assumption that `from` always be for an external Docker image)

* Leave comment on tag tracking being unimplemented (it would consist
  of setting `from:` -> `kind: ImageStream` in an ImageStream, only
  with a different semantics for `name`)