[feature] Also guess the `from:` data structure from the Dockerfile
In the case of an `openshift_imagestream:` with an inline
`dockerfile:` and no other data, we were previously assuming that the
FROM therein would reference an external base image. Support a FROM
line that starts with docker-registry.default.svc:5000 as well.

- Refactor Dockerfile parsing logic into new method `_parse_local_from_lines`
- Call it from `_get_from_struct` method (in addition to existing call site in `_get_build_triggers`)
- In the former, demand exactly one FROM line; otherwise the user should fill in a `from:` section explicitly (we don't want to take a guess according to which of the FROM lines is the “main” one).

Note that “just“ building out of an inline Dockerfile with a FROM line that starts with `docker-registry.default.svc:5000`, doesn't actually work, presumably because the credentials are missing (at any rate, the attempt to  `docker pull` the base image claims that said image doesn't exist, while it actually does). Having a `from:` structure, on the other hand, causes a (null) FROM substitution, and also injects the appropriate credentials into the `docker pull` machinery, so that the pull can succeed.