[fix] Scheduled imports weren't scheduling
`openshift_imagestream`s with a `from:` line and no `source:` are intent on mirroring an image e.g. from the Docker Hub. That's why they set `importPolicy` to `{ "scheduled": True }` without asking. Unfortunately, experimentation (and common sense) suggest that that doesn't have the desired effect (namely, to mirror the upstream image into the ImageStream) when `referencePolicy = { "type": "Source" }` — The latter meaning that Docker builds that rely on such an ImageStream are to bypass the cache.

- Also forcibly set `"referencePolicy": {"type": "Local"}` without asking on `openshift_imagestream`s with a `from:` line and no `source:`