[feature] Ignore diffs that OpenShift brought about by itself
OpenShift has a feature whence image names are updated automatically from ImageStreamTags. (This feature can be turned on using a trigger in a DeploymentConfig, or an `image.openshift.io/triggers` annotation in “vanilla” Kubernetes objects that have pods, such as StatefulSets.)

When this feature is in use, the configuration-as-code will typically contain a symbolic reference to the image (e.g. `u'docker-registry.default.svc:5000/mynamespace/mongo:focal'`), whereas OpenShift will rewrite this pretty much immediately to a resolved image (e.g. u'docker-registry.default.svc:5000/mynamespace/mongo@sha256:4e8a15f....'). This results in a nevergreen situation, plus a gratuitous rollout for objects that are set up to do so on “config change.”

- Grow an `_is_diff_irrelevant` method to filter diffs with
- Detect the case explained above, and return `True`